+++ internal/heartbeat/sysinfo_linux.go
@@ -4,7 +4,7 @@ package heartbeat
 
 import "golang.org/x/sys/unix"
 
-func getTotalMemory() int64 {
+func GetTotalMemory() int64 {
 	var info unix.Sysinfo_t
 	if err := unix.Sysinfo(&info); err != nil {
 		return 0
@@ -12,7 +12,7 @@ func getTotalMemory() int64 {
 	return int64(info.Totalram) * int64(info.Unit)
 }
 
-func getFreeDisk() int64 {
+func GetFreeDisk() int64 {
 	var stat unix.Statfs_t
 	if err := unix.Statfs("/", &stat); err != nil {
 		return 0